lib/group_cpus: optimize inner loop in grp_spread_init_one()
The loop starts from the beginning every time we switch to the next
sibling mask.  This is the Schlemiel the Painter's style of coding because
we know for sure that nmsk is clear up to current CPU, and we can just
continue from the next CPU.

Also, we can do it nicer if leverage the dedicated for_each() iterator.

Link: https://lkml.kernel.org/r/20231207203900.859776-4-yury.norov@gmail.com
Signed-off-by: Yury Norov <yury.norov@gmail.com>
Cc: Andy Shevchenko <andriy.shevchenko@linux.intel.com>
Cc: Ming Lei <ming.lei@redhat.com>
Cc: Rasmus Villemoes <linux@rasmusvillemoes.dk>
Cc: Thomas Gleixner <tglx@linutronix.de>
Signed-off-by: Andrew Morton <akpm@linux-foundation.org>